+++ mujoco_ros/CHANGELOG.md
@@ -1,10 +1,14 @@
 <a name="unreleased"></a>
 ## Unreleased
 
+### Added
+* Compound sensors (for now limited to IMU, Pose, Twist, and Wrench).
+
 ### Fixed
 * Repaired SIGINT handler callback. `C-c` in the roslaunch terminal now shuts down the MuJoCo ROS node instead of escalating to SIGTERM.
 * Added actionlib to the list of mujoco_ros' dependencies.
 * Fixed bug where cutoff was not correctly applied to noisy sensors.
+* Fixed missing FRAMEANGVEL serialization in sensors plugin.
 
 ### Changed
 * replaced `boost::shared_ptr` with `std::shared_ptr` or `std::unique_ptr` wherever possible (ROS 1 fast intra-process message-passing requires boost::shared_ptr).